Added comments explaining parts of code
In case its necessary, if there is an error analysis part of the assignment. We can reference that some of the data had NA values for the things that we tracked. And some even had negative comments. Since you cannot have negative values on youtube in most situations, we just omitted it from our data table in order to simplify things.